% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{approx_control-class}
\alias{approx_control-class}
\title{approx_control S4 class}
\description{
EXPERIMENTAL. See example 7 in \code{vignette("01_additional_examples", package = "geex")}
for usage.
}
\section{Slots}{

\describe{
\item{\code{.FUN}}{a function which approximates an \code{estFUN}.}

\item{\code{.options}}{a list of options passed to \code{.FUN}.}
}}

