% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{basic_control-class}
\alias{basic_control-class}
\title{basic_control S4 class}
\description{
A general class for defining a \code{function}, and the options passed to the
function
}
\section{Slots}{

\describe{
\item{\code{.FUN}}{a function}

\item{\code{.options}}{a list of options passed to \code{.FUN}}
}}

\seealso{
\code{\link{root_control-class}}, \code{\link{deriv_control-class}}
\code{\link{approx_control-class}}
}
