% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_eeFUN_funs.R
\name{binomial_integrand}
\alias{binomial_integrand}
\title{Objective Function for Logistic-Normal Likelihood}
\usage{
binomial_integrand(b, response, xmatrix, parms, linkinv)
}
\arguments{
\item{b}{the random effect to integrate over}

\item{response}{vector of response values}

\item{xmatrix}{the matrix of covariates}

\item{parms}{vector of parameters}

\item{linkinv}{inverse link function}
}
\description{
Objective Function for Logistic-Normal Likelihood
}
