% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_array}
\alias{check_array}
\title{Check an array object}
\usage{
check_array(object)
}
\arguments{
\item{object}{the object to check whether it is an array}
}
\value{
an array - either the orginal object or the given object converted
to an array
}
\description{
Handles the case where there is a single estimating equation. This function
assumes that the object
}
