% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_funs.R
\name{check_corrections}
\alias{check_corrections}
\title{Check that a list of corrections conforms to geex standards}
\usage{
check_corrections(corrections)
}
\arguments{
\item{corrections}{a list of small sample corrections to perform}
}
\description{
Check that a list of corrections conforms to geex standards
}
