% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_eeFUN}
\alias{check_eeFUN}
\title{Check an eeFUN object}
\usage{
check_eeFUN(geex_list)
}
\arguments{
\item{geex_list}{a list of \code{eeFUN}, \code{splitdt}, \code{inner_eeargs}, and
\code{outer_eeargs}}
}
\description{
Checks that eeFUN returns a function
}
