% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{check_GFUN}
\alias{check_GFUN}
\title{Check value of estimating eqations}
\usage{
check_GFUN(geex_list, theta)
}
\arguments{
\item{geex_list}{a list with at least \code{splitdt}, \code{eeFUN}, and \code{outer_eeargs} (which can be \code{NULL})}

\item{theta}{parameter estimates to use in evaluating the estimating equations.}
}
\description{
Checks that the value of Gm = sum_i psi(O_i, theta) is close to zero when
evaluated at estimated parameters.
}
