% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{coef,geex-method}
\alias{coef,geex-method}
\title{Gets the parameter estimates matrix from a geex object}
\usage{
\S4method{coef}{geex}(object)
}
\description{
Gets the parameter estimates matrix from a geex object
}
