% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computations.R
\name{compute_eeroot}
\alias{compute_eeroot}
\title{Compute roots for a set of estimating equations}
\usage{
compute_eeroot(geex_list, rootFUN = rootSolve::multiroot,
  rootFUN_control = NULL, approxFUN = NULL, approxFUN_control = NULL)
}
\arguments{
\item{geex_list}{a list containing \code{splitdt} (a \code{data.frame} that
has been \code{\link[base]{split}} by the grouping variable) and \code{eeFUN}
(see \code{\link{estimate_equations}}). Note that the function that is
returned by \code{eeFUN} must take \code{theta} as its first argument, where
\code{theta} represents the parameters.}

\item{rootFUN}{the function used to find roots of the estimating equations.
Defaults to \code{\link[rootSolve]{multiroot}}.}

\item{rootFUN_control}{a list of options to be passed to the \code{rootsolver}
function}

\item{approxFUN}{a function that approximates the inner function of \code{eeFUN}}

\item{approxFUN_control}{arguments passed to \code{approxFUN}}
}
\value{
the output of the \code{rootsolver} function
}
\description{
Compute roots for a set of estimating equations
}
