% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_funs.R
\name{compute_pairwise_sum_of_list}
\alias{compute_pairwise_sum_of_list}
\title{Compute the sum of  a list of matrices to sum}
\usage{
compute_pairwise_sum_of_list(.l, .w = NULL, .wFUN = NULL, ...)
}
\arguments{
\item{.l}{a list of matrices}

\item{.w}{a numeric vector of weights}

\item{.wFUN}{a function of \code{i}, \code{j}, and (optionally) additional
arguments}

\item{...}{additional arguments passed to \code{.wFUN}

Either \code{.w} or \code{.wFUN} must be specified but not both.}
}
\description{
Compute the sum of  a list of matrices to sum
}
\keyword{internal}
