% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_sigma}
\alias{compute_sigma}
\title{Compute covariance matrix for set of estimating equations}
\usage{
compute_sigma(matrices, corrections = NULL)
}
\description{
Compute covariance matrix for set of estimating equations
}

