% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_funs.R
\name{compute_sum_of_list}
\alias{compute_sum_of_list}
\title{Compute the sum of  a list of matrices to sum}
\usage{
compute_sum_of_list(.l, .w = numeric(0))
}
\arguments{
\item{.l}{a list of matrices}

\item{.w}{a numeric vector of weights}
}
\description{
Compute the sum of  a list of matrices to sum
}
