% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{control-class}
\alias{control-class}
\title{control S4 class}
\description{
control S4 class
}
\section{Slots}{

\describe{
\item{\code{.FUN}}{a function}

\item{\code{.options}}{a list of options passed to \code{.FUN}}
}}

