% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_by_funs.R
\name{correct_by_fay_bias}
\alias{correct_by_fay_bias}
\title{Estimate Fay's bias correction}
\usage{
correct_by_fay_bias(A_i, A, B_i, B, b = 0.75)
}
\arguments{
\item{A_i}{a list of bread matrices per group}

\item{A}{the outer "bread" matrix}

\item{B_i}{a list of meat matrices per group}

\item{B}{the inner "meat" matrix}

\item{b}{a numeric value < 1. Defaults to 0.75 as in Fay.}
}
\value{
a corrected covariance matrix
}
\description{
Computes the bias corrected sandwich covariance matrix described in Fay and
Graubard (2001).
}
\references{
Fay, M. P., & Graubard, B. I. (2001). Small-Sample adjustments for
Wald-type tests using sandwich estimators. Biometrics, 57(4), 1198-1206
}
