% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{correct_control-class}
\alias{correct_control-class}
\title{correct_control S4 class}
\description{
correct_control S4 class
}
\section{Slots}{

\describe{
\item{\code{.FUN}}{a function which "corrects" a \code{\linkS4class{sandwich_components}}
object. Usually a small-sample correction}

\item{\code{.options}}{a list of options passed to \code{.FUN}.}
}}

