% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_funs.R
\name{create_GFUN}
\alias{create_GFUN}
\title{Creates a function that sums over psi functions}
\usage{
create_GFUN(.basis)
}
\arguments{
\item{.basis}{an object of class \code{\linkS4class{m_estimation_basis}}}
}
\description{
From a list of \eqn{\psi(O_i, \theta)}{\psi(O_i, \theta)} for i = 1, ..., m,
creates \eqn{G_m = \sum_i \psi(O_i, \theta)}{G_m = \sum_i \psi(O_i, \theta)},
called \code{GFUN}. Here, \eqn{\psi(O_i, \theta)}{\psi(O_i, \theta)} is the
*inner* part of an \code{estFUN}, in that the data is fixed and \eqn{G_m}{G_m}
is a function of \eqn{\theta)}{\theta}.
}
