% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computations.R
\name{create_psi}
\alias{create_psi}
\title{Creates list of psi functions}
\usage{
create_psi(splitdt, eeFUN, approxFUN = NULL, approxFUN_control = NULL,
  outer_eeargs = NULL)
}
\arguments{
\item{splitdt}{list of dataframes with data per unit}

\item{eeFUN}{the estimating equation function}

\item{approxFUN}{a function that approximates the inner function of \code{eeFUN}}

\item{approxFUN_control}{arguments passed to \code{approxFUN}}
}
\description{
Creates list of psi functions
}

