% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_funs.R
\name{create_psiFUN_list}
\alias{create_psiFUN_list}
\title{Creates list of psi functions}
\usage{
create_psiFUN_list(object, ...)
}
\arguments{
\item{.basis}{an object of class \code{\linkS4class{m_estimation_basis}}}
}
\value{
the \code{.basis} with the \code{.psiFUN_list} slot populated.
}
\description{
Creates the estimating function (\eqn{\psi(O_i, \theta)}{\psi(O_i, \theta)})
for each unit. That is, this function evaluates the outer function in
\code{estFUN} for each independent unit and a returns the inner function in
\code{estFUN}.
}
