% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{deriv_control-class}
\alias{deriv_control-class}
\title{deriv_control S4 class}
\description{
deriv_control S4 class
}
\section{Slots}{

\describe{
\item{\code{.FUN}}{a function which computes a numerical derivation. This functions
first argument must the function on which the derivative is being compute.
Defaults to \code{\link[numDeriv]{jacobian}}.}

\item{\code{.options}}{a list of options passed to \code{.FUN}. Defaults to
\code{list(method = 'Richardson')}}
}}

