% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_by_funs.R
\name{df_correction_1}
\alias{df_correction_1}
\title{Estimate Fay's degrees of freedom corrections 1}
\usage{
df_correction_1(A_d, C)
}
\arguments{
\item{A_d}{see \code{\link{df_correction_prep}}}

\item{C}{see \code{\link{df_correction_prep}}}
}
\description{
Estimate Fay's degrees of freedom corrections 1
}
