% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrections.R
\name{df_correction_prep}
\alias{df_correction_prep}
\title{Preparations for Fay's degrees of freedom corrections}
\usage{
df_correction_prep(m, L, A, A_i, H_i)
}
\description{
Preparations for Fay's degrees of freedom corrections
}

