% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_funs.R
\name{diagnose_roots}
\alias{diagnose_roots}
\title{Diagnose roots of estimating equations}
\usage{
diagnose_roots(geex_list, theta)
}
\arguments{
\item{geex_list}{a list with at least \code{splitdt}, \code{eeFUN}, and \code{outer_eeargs} (which can be \code{NULL})}

\item{theta}{parameter estimates to use in evaluating the estimating equations.}
}
\value{
a numeric vector
}
\description{
Computes the value of \deqn{G_m = sum_i psi(O_i, \hat{\theta})}{G_m = sum_i psi(O_i, theta)}, i.e., the estimating
equations at \code{theta}. Used to verify that \eqn{G_m = 0}{G_m = 0} (or close to 0).
}
