% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_by_funs.R
\name{estimate_df}
\alias{estimate_df}
\title{Estimate Fay's degrees of freedom corrections}
\usage{
estimate_df(A, C)
}
\arguments{
\item{A}{the outer "bread" matrix}

\item{C}{same as the C matrix in the Fay (2001) notation Section 2.3}
}
\description{
Estimate Fay's degrees of freedom corrections
}
