% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_funs.R
\name{estimate_GFUN_roots}
\alias{estimate_GFUN_roots}
\title{Estimates roots for a set of estimating equations}
\usage{
estimate_GFUN_roots(basis, rootFUN = rootSolve::multiroot,
  rootFUN_control = NULL, approxFUN = NULL, approxFUN_control = NULL)
}
\arguments{
\item{basis}{an object of class \code{\linkS4class{m_estimation_basis}}}

\item{rootFUN}{the function used to find roots of the estimating equations.
Defaults to \code{\link[rootSolve]{multiroot}}.}

\item{rootFUN_control}{a list of options to be passed to the \code{rootsolver}
function}

\item{approxFUN}{a function that approximates the inner function of \code{eeFUN}.
(EXPERIMENTAL).}

\item{approxFUN_control}{arguments passed to \code{approxFUN}}
}
\value{
the output of the \code{rootFUN} function
}
\description{
Using the \code{rootFUN} specified by the user (defaults to \code{\link[rootSolve]{multiroot}}),
this function estimates the roots of the equations:
\deqn{G_m = sum_i psi(O_i, \hat{\theta}) = 0}{G_m = sum_i psi(O_i, theta) = 0}
}
\details{
This is primilary an internal function used within \code{\link{estimate_equations}},
but it is exported for use in debugging and development.

For an example of how to use a different \code{rootFUN},
see the root solver vignette, \code{vignette('geex_root_solvers', package = 'geex')}.
}
