% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{estimating_function-class}
\alias{estimating_function-class}
\title{estimating_function S4 class}
\description{
estimating_function S4 class
}
\section{Slots}{

\describe{
\item{\code{.estFUN}}{the estimating function.}

\item{\code{.outer_args}}{a named \code{list} of arguments passed to the outer
function of \code{estFUN}. Should *not* include the \code{data} argument.}

\item{\code{.inner_args}}{a named \code{list} of arguments passed to the inner
function of \code{estFUN}. Should *not* include the \code{theta} argument.}
}}

