% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_by_funs.R
\name{fay_bias_correction_partial}
\alias{fay_bias_correction_partial}
\title{Compute the matrices necessary for Fay's bias correction}
\usage{
fay_bias_correction_partial(A, A_i, B, B_i, b)
}
\arguments{
\item{A}{the outer "bread" matrix}

\item{A_i}{a list of bread matrices per group}

\item{B}{the inner "meat" matrix}

\item{B_i}{a list of meat matrices per group}

\item{b}{a numeric value < 1. Defaults to 0.75 as in Fay.}
}
\value{
a list with
\itemize{
\item H_i - the H_i matrix
\item Bbc - the 'bias corrected' meat matrix
}
}
\description{
Compute the matrices necessary for Fay's bias correction
}
