% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{geex-class}
\alias{geex-class}
\title{geex S4 class}
\description{
geex S4 class
}
\section{Slots}{

\describe{
\item{\code{call}}{the \code{m_estimate} call}

\item{\code{basis}}{a \code{\linkS4class{m_estimation_basis}} object}

\item{\code{rootFUN_results}}{the results of call to the root finding algorithm function}

\item{\code{sandwich_components}}{a \code{\linkS4class{sandwich_components}} object}

\item{\code{GFUN}}{the \code{function} of which the roots are computed.}

\item{\code{corrections}}{a \code{list} of correction performed on \code{sandwich_components}}

\item{\code{estimates}}{a \code{numeric} vector of parameter estimates}

\item{\code{vcov}}{the empirical sandwich variance \code{matrix}}
}}

