% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{geex-summary-class}
\alias{geex-summary-class}
\title{geex summary object}
\description{
geex summary object
}
\section{Slots}{

\describe{
\item{\code{estFUN}}{a \code{estimating-function}}

\item{\code{outer_args}}{the \code{list} arguments passed to the \code{m_estimate} call}

\item{\code{inner_args}}{the \code{list} arguments passed to the \code{m_estimate} call}

\item{\code{data}}{the \code{data.frame} passed to the \code{m_estimate} call}

\item{\code{weights}}{the \code{weights} passed to the \code{m_estimate} call}

\item{\code{nobs}}{the number of observational units used to compute the M-estimator}

\item{\code{units}}{the name of the variable identifying the observational units}

\item{\code{corrections}}{a \code{list} of correction performed on \code{sandwich_components}}

\item{\code{estimates}}{a \code{numeric} vector of parameter estimates}

\item{\code{vcov}}{the empirical sandwich variance \code{matrix}}
}}

