% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{geex_control-class}
\alias{geex_control-class}
\title{geex_control S4 class}
\description{
An object which control all the \code{\linkS4class{basic_control}} objects
necessary to perform M-estimation
}
\section{Slots}{

\describe{
\item{\code{.approx}}{an \code{\linkS4class{approx_control}} object}

\item{\code{.root}}{a \code{\linkS4class{root_control}} object}

\item{\code{.deriv}}{a \code{\linkS4class{deriv_control}} object}
}}

