% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{get_corrections}
\alias{get_corrections}
\alias{get_corrections,geex-method}
\alias{get_corrections,m_estimation_basis,m_estimation_basis-method}
\title{Gets the corrections from a geex object}
\usage{
get_corrections(object, ...)

\S4method{get_corrections}{geex}(object)
}
\arguments{
\item{object}{a \code{\linkS4class{geex}} object}

\item{...}{arguments passed to other methods}
}
\description{
Gets the corrections from a geex object
}
