% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_design_matrix}
\alias{get_design_matrix}
\title{Get a matrix of fixed effects from a model object}
\usage{
get_design_matrix(rhs_formula, data)
}
\arguments{
\item{rhs_formula}{the right hand side of a model formula}

\item{data}{the data from which to extract the matrix}
}
\description{
Get a matrix of fixed effects from a model object
}

