% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_response}
\alias{get_response}
\title{Get a vector of responses from a model object}
\usage{
get_response(formula, data)
}
\arguments{
\item{formula}{model formula}

\item{data}{data.frame from which to extract the vector of responses}
}
\description{
Get a vector of responses from a model object
}

