% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_response_formula}
\alias{get_response_formula}
\title{Get the LHS formula from a model object}
\usage{
get_response_formula(model)
}
\arguments{
\item{model}{a model object such as \code{lm}, \code{glm}, \code{merMod}}
}
\description{
Get the LHS formula from a model object
}
