% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{grab_approx_options}
\alias{grab_approx_options}
\title{grab_approx_options method}
\usage{
grab_approx_options(object, ...)
}
\description{
Extracts the \code{.approx@.options} slot from a \code{\linkS4class{geex_control}} object
}
