% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{grab_approxFUN}
\alias{grab_approxFUN}
\title{grab_approxFUN method}
\usage{
grab_approxFUN(object, ...)
}
\description{
Extracts the \code{.approx@.FUN} slot from a \code{\linkS4class{geex_control}} object
}
