% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{grab_basis_data}
\alias{grab_basis_data}
\title{grab_basis_data generic}
\usage{
grab_basis_data(object, ...)
}
\description{
Grabs the \code{.data} from an \code{\linkS4class{m_estimation_basis}} object
}
