% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{grab_bread}
\alias{grab_bread}
\title{Grabs the .A (bread matrix) slot}
\usage{
grab_bread(object)
}
\arguments{
\item{object}{a \code{\linkS4class{sandwich_components}} object}
}
\description{
Grabs the .A (bread matrix) slot
}
