% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{grab_bread_list}
\alias{grab_bread_list}
\alias{grab_bread_list,sandwich_components-method}
\alias{grab_bread_list,sandwich_components,sandwich_components-method}
\title{Gets the .A_i (list of bread matrices) slot}
\usage{
grab_bread_list(object)

\S4method{grab_bread_list}{sandwich_components}(object)
}
\arguments{
\item{object}{a \code{\linkS4class{sandwich_components}} object}
}
\description{
Gets the .A_i (list of bread matrices) slot
}
