% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab.R
\name{grab_design_matrix}
\alias{grab_design_matrix}
\title{Grab a matrix of fixed effects from a model object}
\usage{
grab_design_matrix(data, rhs_formula)
}
\arguments{
\item{data}{the data from which to extract the matrix}

\item{rhs_formula}{the right hand side of a model formula}
}
\description{
Grab a matrix of fixed effects from a model object
}
