% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_functions.R
\name{grab_eeFUN.geeglm}
\alias{grab_eeFUN.geeglm}
\title{geeglm Estimating Equations}
\usage{
\method{grab_eeFUN}{geeglm}(model, data, ...)
}
\arguments{
\item{model}{a model object object}

\item{data}{data with which to create the estimating equation function}

\item{...}{passed to methods}
}
\description{
Create estimating equation function from a \code{geeglm} object
}
