% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_eeFUN_funs.R
\name{grab_eeFUN.glm}
\alias{grab_eeFUN.glm}
\title{glm Estimating Equations}
\usage{
\method{grab_eeFUN}{glm}(model, data, weights = 1, ...)
}
\arguments{
\item{model}{a model object object}

\item{data}{data with which to create the estimating equation function}

\item{weights}{a scalar or vector of weight values}

\item{...}{passed to methods}
}
\description{
Create estimating equation function from a \code{glm} object
}
