% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_functions.R
\name{grab_eeFUN}
\alias{grab_eeFUN}
\title{Make Estimating Equation functions}
\usage{
grab_eeFUN(model, data, ...)
}
\arguments{
\item{model}{a model object object}

\item{data}{data with which to create the estimating equation function}

\item{...}{passed to methods}
}
\description{
Converts a model object into
}
