% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_estFUN_funs.R
\docType{methods}
\name{grab_estFUN,merMod-method}
\alias{grab_estFUN,merMod-method}
\title{Grab estimating functions from a merMod object}
\usage{
\S4method{grab_estFUN}{merMod}(object, data, numderiv_opts = NULL, ...)
}
\arguments{
\item{numderiv_opts}{a list of argument passed to \code{numDeriv::grad}}
}
\description{
Create estimating equation function from a \code{merMod} object
}
