% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{grab_estFUN}
\alias{grab_estFUN}
\alias{grab_estFUN,estimating_function-method}
\alias{grab_estFUN,estimating_function,estimating_function-method}
\title{Grab estimating functions from a model object}
\usage{
grab_estFUN(object)

\S4method{grab_estFUN}{estimating_function}(object)
}
\arguments{
\item{object}{a \code{\linkS4class{estimating_function}} object}
}
\description{
Grab estimating functions from a model object
}
