% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_estFUN_funs.R
\name{grab_estFUN.geeglm}
\alias{grab_estFUN.geeglm}
\title{Grab estimating functions from a geeglm object}
\usage{
grab_estFUN.geeglm(object, data, ...)
}
\description{
Create estimating equation function from a \code{geeglm} object
}
