% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_estFUN_funs.R
\name{grab_estFUN.glm}
\alias{grab_estFUN.glm}
\title{Grab estimating functions from a glm object}
\usage{
grab_estFUN.glm(object, data, weights = 1, ...)
}
\arguments{
\item{weights}{a scalar or vector of weight values}
}
\description{
Create estimating equation function from a \code{glm} object
}
