% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_estFUN_funs.R
\name{grab_estFUN.merMod}
\alias{grab_estFUN.merMod}
\title{Grab estimating functions from a merMod object}
\usage{
grab_estFUN.merMod(object, data, numderiv_opts = NULL, ...)
}
\arguments{
\item{numderiv_opts}{a list of argument passed to \code{numDeriv::grad}}
}
\description{
Create estimating equation function from a \code{merMod} object
}
