% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_estFUN_funs.R
\name{grab_estFUN}
\alias{grab_estFUN}
\alias{grab_estFUN}
\title{Grab estimating functions from a model object}
\usage{
grab_estFUN(object, ...)

grab_estFUN(object, ...)
}
\description{
Grab estimating functions from a model object

Grabs the \code{.estFUN} from an \code{\linkS4class{m_estimation_basis}} object
}
