% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_funs.R
\name{grab_fixed_formula}
\alias{grab_fixed_formula}
\title{Grab the RHS formula from a model object}
\usage{
grab_fixed_formula(model)
}
\arguments{
\item{model}{a model object such as \code{lm}, \code{glm}, \code{merMod}}
}
\description{
Grab the RHS formula from a model object
}
