% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{grab_FUN,geex_control-method}
\alias{grab_FUN,geex_control-method}
\title{Extracts the \code{.approx@.FUN} slot from a \code{\linkS4class{geex_control}} object}
\usage{
\S4method{grab_FUN}{geex_control}(object, slot)
}
\arguments{
\item{object}{a \code{\linkS4class{geex_control}} object}

\item{slot}{name of the slot from which to grab the function. One of "deriv",
"approx", or "root"}
}
\description{
Extracts the \code{.approx@.FUN} slot from a \code{\linkS4class{geex_control}} object
}
