% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{grab_FUN}
\alias{grab_FUN}
\title{Extracts the \code{.FUN} slot from a \code{\linkS4class{basic_control}} object}
\usage{
grab_FUN(object, ...)
}
\arguments{
\item{object}{a \code{\linkS4class{basic_control}} object}
}
\description{
Extracts the \code{.FUN} slot from a \code{\linkS4class{basic_control}} object
}
