% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{grab_GFUN}
\alias{grab_GFUN}
\alias{grab_GFUN,m_estimation_basis-method}
\alias{grab_GFUN,m_estimation_basis,m_estimation_basis-method}
\title{Gets the .psi_list slot in a m_estimation_basis}
\usage{
grab_GFUN(object)

\S4method{grab_GFUN}{m_estimation_basis}(object)
}
\arguments{
\item{object}{a \code{\linkS4class{m_estimation_basis}} object}
}
\description{
Gets the .psi_list slot in a m_estimation_basis
}
