% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{grab_GFUN}
\alias{grab_GFUN}
\title{Gets the .psi_list slot in a m_estimation_basis}
\usage{
grab_GFUN(object)
}
\arguments{
\item{object}{a \code{\linkS4class{m_estimation_basis}} object}
}
\description{
Gets the .psi_list slot in a m_estimation_basis
}
