% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{grab_meat}
\alias{grab_meat}
\title{Gets the .B (meat matrix) slot}
\usage{
grab_meat(object)
}
\arguments{
\item{object}{a \code{\linkS4class{sandwich_components}} object}
}
\description{
Gets the .B (meat matrix) slot
}
