% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{grab_meat_list}
\alias{grab_meat_list}
\alias{grab_meat_list,sandwich_components-method}
\alias{grab_meat_list,sandwich_components,sandwich_components-method}
\title{Gets the .B_i (list of bread matrices) slot}
\usage{
grab_meat_list(object)

\S4method{grab_meat_list}{sandwich_components}(object)
}
\arguments{
\item{object}{a \code{\linkS4class{sandwich_components}} object}
}
\description{
Gets the .B_i (list of bread matrices) slot
}
