% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{grab_options}
\alias{grab_options}
\alias{grab_options,basic_control-method}
\alias{grab_options,basic_control,basic_control-method}
\alias{grab_options,geex_control-method}
\alias{grab_options,geex_control,geex_control-method}
\title{Extracts the \code{.options} slot from a \code{\linkS4class{basic_control}} object}
\usage{
grab_options(object, ...)

\S4method{grab_options}{basic_control}(object)

\S4method{grab_options}{geex_control}(object, slot)
}
\arguments{
\item{object}{a \code{\linkS4class{basic_control}} object}

\item{...}{additional arguments passed to other methods}

\item{slot}{name of the slot from which to grab the function. One of "deriv",
"approx", or "root"}
}
\description{
Extracts the \code{.options} slot from a \code{\linkS4class{basic_control}} object

Extracts options from a \code{\linkS4class{geex_control}} object
}
