% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab.R
\name{grab_response}
\alias{grab_response}
\title{Grab a vector of responses from a model object}
\usage{
grab_response(data, formula)
}
\arguments{
\item{data}{data.frame from which to extract the vector of responses}

\item{formula}{model formula}
}
\description{
Grab a vector of responses from a model object
}
