% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab.R
\name{grab_response_formula}
\alias{grab_response_formula}
\title{Grab the LHS formula from a model object}
\usage{
grab_response_formula(model)
}
\arguments{
\item{model}{a model object such as \code{lm}, \code{glm}, \code{merMod}}
}
\description{
Grab the LHS formula from a model object
}
