% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{initialize,m_estimation_basis-method}
\alias{initialize,m_estimation_basis-method}
\title{Initialize a m_estimation_basis object}
\usage{
\S4method{initialize}{m_estimation_basis}(.Object, ...)
}
\description{
# @export
}
