% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_functions.R
\name{make_eefun.glm}
\alias{make_eefun.glm}
\title{glm Estimating Equations}
\usage{
\method{make_eefun}{glm}(model, data, weights = 1, ...)
}
\arguments{
\item{model}{a model object object}

\item{data}{data with which to create the estimating equation function}

\item{weights}{a scalar or vector of weight values}

\item{...}{passed to methods}
}
\description{
Create estimating equation function from a \code{glm} object
}

