% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_functions.R
\name{make_eefun.merMod}
\alias{make_eefun.merMod}
\title{glmer Estimating Equations}
\usage{
\method{make_eefun}{merMod}(object, data, numderiv_opts = NULL)
}
\arguments{
\item{data}{data with which to create the estimating equation function}
}
\description{
Create estimating equation function from a \code{merMod} object
}

