% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_functions.R
\name{make_eefun}
\alias{make_eefun}
\title{Make Estimating Equation functions}
\usage{
make_eefun(model, data, ...)
}
\arguments{
\item{model}{a model object object}

\item{data}{data with which to create the estimating equation function}

\item{...}{passed to methods}
}
\description{
Converts a model object into
}
