% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_functions.R
\name{objFun_glmerMod_binomial}
\alias{objFun_glmerMod_binomial}
\title{glmer Objective Function for Logistic-Normal Likelihood}
\usage{
objFun_glmerMod_binomial(parms, response, xmatrix, linkinv)
}
\arguments{
\item{parms}{vector of parameters}

\item{response}{vector of response values}

\item{xmatrix}{the matrix of covariates}

\item{linkinv}{inverse link function}
}
\description{
glmer Objective Function for Logistic-Normal Likelihood
}

