% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{root_control-class}
\alias{root_control-class}
\title{root_control S4 class}
\description{
root_control S4 class
}
\section{Slots}{

\describe{
\item{\code{.FUN}}{a root finding function whose first argument must be named \code{f}.}

\item{\code{.options}}{a list of options passed to \code{.FUN}.}

\item{\code{.object_name}}{a character string identifying the object containing the
roots in the output of \code{.FUN}.}
}}

