% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{roots}
\alias{roots}
\alias{roots,geex-method}
\alias{roots,geex,geex-method}
\title{Gets the parameter estimates matrix from a geex object}
\usage{
roots(object, ...)

\S4method{roots}{geex}(object)
}
\arguments{
\item{object}{a \code{\linkS4class{geex}} object}

\item{...}{arguments passed to other methods}
}
\description{
Gets the parameter estimates matrix from a geex object
}
