% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{sandwich_components-class}
\alias{sandwich_components-class}
\title{sandwich_components S4 class}
\description{
sandwich_components S4 class
}
\section{Slots}{

\describe{
\item{\code{.A}}{the "bread" matrix}

\item{\code{.A_i}}{a list of "bread" matrices per unit}

\item{\code{.B}}{the "meat" matrix}

\item{\code{.B_i}}{a list of "meat" matrices per unit}
}}

