% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{set_control<-}
\alias{set_control<-}
\title{approxOPTS generic}
\usage{
set_control(object, slot) <- value
}
\description{
Extracts the \code{.approx@.options} slot from a \code{\linkS4class{geex_control}} object
}
