% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{set_GFUN<-}
\alias{set_GFUN<-}
\title{Sets the .psi_list slot in a m_estimation_basis}
\usage{
set_GFUN(object) <- value
}
\arguments{
\item{object}{a \code{\linkS4class{m_estimation_basis}} object}

\item{value}{a \code{function} created \code{\link{create_GFUN}}}
}
\description{
Sets the .psi_list slot in a m_estimation_basis
}
