% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{set_psiFUN_list<-}
\alias{set_psiFUN_list<-}
\title{Sets the .psi_list slot in a m_estimation_basis}
\usage{
set_psiFUN_list(object) <- value
}
\description{
Sets the .psi_list slot in a m_estimation_basis
}
