% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{setup_approx_control}
\alias{setup_approx_control}
\title{Setup an approx _control object}
\usage{
setup_approx_control(FUN, ...)
}
\arguments{
\item{FUN}{a function}

\item{...}{arguments passed to \code{FUN}}
}
\value{
a \code{\linkS4class{approx_control}} object
}
\description{
Setup an approx _control object
}
