% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{setup_control}
\alias{setup_control}
\title{Setup a basic_control object}
\usage{
setup_control(type, FUN, ...)
}
\arguments{
\item{type}{one of \code{c("deriv", "approx", "root")}}

\item{FUN}{a function}

\item{...}{arguments passed to \code{FUN}}
}
\value{
a \code{\linkS4class{basic_control}} object
}
\description{
Setup a basic_control object
}
