% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{setup_deriv_control}
\alias{setup_deriv_control}
\title{Setup a deriv_control object}
\usage{
setup_deriv_control(FUN, ...)
}
\arguments{
\item{FUN}{a function}

\item{...}{arguments passed to \code{FUN}}
}
\value{
a \code{\linkS4class{deriv_control}} object
}
\description{
Setup a deriv_control object
}
