% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{setup_root_control}
\alias{setup_root_control}
\title{Setup a root_control object}
\usage{
setup_root_control(FUN, roots_name, ...)
}
\arguments{
\item{FUN}{a root finding function whose first argument must be named \code{f}.
Defaults to \code{\link[rootSolve]{multiroot}}}

\item{roots_name}{a character string identifying the object containing the
roots in the output of \code{FUN}. Defaults to \code{"root"}}

\item{...}{arguments passed to \code{FUN}}
}
\value{
a \code{\linkS4class{root_control}} object
}
\description{
Setup a root_control object
}
\examples{
# Setup the default
setup_root_control(start = c(3, 5, 6))

# Also setup the default
setup_root_control(FUN = rootSolve::multiroot,
                   start = c(3, 5, 6))

# Or use uniroot()
setup_root_control(FUN = stats::uniroot,
                   interval = c(0, 1))
}
