% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{setup_root_solver}
\alias{setup_root_solver}
\title{Setup a root_control object}
\usage{
setup_root_solver(FUN, roots_name, ...)
}
\arguments{
\item{FUN}{a root finding function whose first argument must be named \code{f}.}

\item{roots_name}{a character string identifying the object containing the
roots in the output of \code{FUN}.}

\item{...}{arguments passed to \code{FUN}}
}
\value{
a \code{\linkS4class{root_control}} object
}
\description{
Setup a root_control object
}
