% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{setup_rootFUN}
\alias{setup_rootFUN}
\title{Setup a root_control object}
\usage{
setup_rootFUN(rootFUN, roots_name, ...)
}
\arguments{
\item{rootFUN}{a root finding function whose first argument must be named \code{f}.}

\item{roots_name}{a character string identifying the object containing the
roots in the output of \code{.FUN}.}

\item{...}{arguments passed to \code{rootFUN}}
}
\description{
Setup a root_control object
}
