% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{show,m_estimation_basis-method}
\alias{show,m_estimation_basis-method}
\title{Shows the m_estimation_basis}
\usage{
\S4method{show}{m_estimation_basis}(object)
}
\description{
Shows the m_estimation_basis
}
