% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{show,sandwich_components-method}
\alias{show,sandwich_components-method}
\title{Shows the sandwich_components S4 class}
\usage{
\S4method{show}{sandwich_components}(object)
}
\arguments{
\item{object}{a \code{\linkS4class{sandwich_components}} object}
}
\description{
Shows the sandwich_components S4 class
}
