% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{vcov,geex-method}
\alias{vcov,geex-method}
\title{Gets the variance-covariance matrix from a geex object}
\usage{
\S4method{vcov}{geex}(object)
}
\description{
Gets the variance-covariance matrix from a geex object
}
