% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{weights,geex-method}
\alias{weights,geex-method}
\alias{weights,geex,geex-method}
\alias{weights,geex_summary-method}
\alias{weights,geex_summary,geex_summary-method}
\title{Extract Model weights}
\usage{
\S4method{weights}{geex}(object)

\S4method{weights}{geex_summary}(object)
}
\arguments{
\item{object}{a \code{\linkS4class{geex}} object}
}
\description{
Extract Model weights
}
