% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrections.R
\name{check_corrections}
\alias{check_corrections}
\title{Check that a list of corrections conforms to geex standards}
\usage{
check_corrections(corrections)
}
\arguments{
\item{corrections}{a list with one sublist for each correction to make. The
sublist must contain at least \code{correctFUN} and optionally additional
arguments in \code{correctFUN_control}.}
}
\description{
Check that a list of corrections conforms to geex standards
}
