% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computations.R
\name{compute_eeroot}
\alias{compute_eeroot}
\title{Compute roots for a set of estimating equations}
\usage{
compute_eeroot(geex_list, start = NULL, rootFUN = rootSolve::multiroot,
  rootFUN_control = NULL, ...)
}
\arguments{
\item{geex_list}{a list containing \code{splitdt} (a \code{data.frame} that
has been \code{\link[base]{split}} by the grouping variable) and \code{eeFUN}
(see \code{\link{estimate_equations}}). Note that the function that is
returned by \code{eeFUN} must take \code{theta} as its first argument, where
\code{theta} represents the parameters.}

\item{start}{vector with length of the number of parameters to find. Passed to
\code{\link[rootSolve]{multiroot}} if not NULL. Defaults to NULL.}

\item{rootFUN}{the function used to find roots of the estimating equations.
Defaults to \code{\link[rootSolve]{multiroot}}.}

\item{rootFUN_control}{a list of options to be passed to the \code{rootsolver}
function}

\item{...}{additional arguments passed to \code{geex_list$eeFUN}}
}
\value{
the output of the \code{rootsolver} function
}
\description{
Compute roots for a set of estimating equations
}
