% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_matrices}
\alias{compute_matrices}
\title{Compute component matrices for covariance matrix for a set of
estimating equations}
\usage{
compute_matrices(obj, contrast = NULL, theta, corrections = NULL,
  correction_options = list(), numDeriv_options = list(method =
  "Richardson"), ...)
}
\description{
Compute component matrices for covariance matrix for a set of
estimating equations
}

