% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computations.R
\name{compute_matrices}
\alias{compute_matrices}
\title{Compute component matrices of the empirical sandwich covariance estimator}
\usage{
compute_matrices(geex_list, theta, derivFUN = numDeriv::jacobian,
  derivFUN_control = list(method = "Richardson"), approxFUN = NULL,
  approxFUN_control = NULL)
}
\arguments{
\item{geex_list}{a list containing \code{splitdt} (a \code{data.frame} that
has been \code{\link[base]{split}} by the grouping variable) and \code{eeFUN}
(see \code{\link{estimate_equations}})}

\item{theta}{vector of parameter estimates (i.e. estimated roots) passed to \code{geex_list$eeFUN}.}

\item{derivFUN}{the function used to take derivatives of the estimating equation functions.
Defaults to \code{\link[numDeriv]{jacobian}}.}

\item{derivFUN_control}{a list of options passed to \code{\link[numDeriv]{jacobian}}
(or the \code{derivFUN} function).}

\item{approxFUN}{a function that approximates the inner function of \code{eeFUN}.
(EXPERIMENTAL).}

\item{approxFUN_control}{arguments passed to \code{approxFUN}}
}
\value{
a list with
\itemize{
\item A - the 'bread' matrix
\item B - the 'meat' matrix
\item A_i - a list of 'bread' matrices for each group
\item B_i - a list of 'meat' matrices for each group
}
}
\description{
For a given set of estimating equations computes the 'meat' (\eqn{B_m}{B_m} in Stefanski and Boos notation)
and 'bread' (\eqn{A_m}{A_m} in Stefanski and Boos notation) matrices necessary to compute the covariance matrix.
}
\details{
For a set of estimating equations (\eqn{\sum_i \psi(O_i, \theta) = 0}{sum_i \psi(O_i, \theta) = 0}),
this function computes:

\deqn{A_i =  \partial \psi(O_i, \theta)/\partial \theta}{A_i =  partial \psi(O_i, theta)/\partial \theta}

\deqn{A =  \sum_i A_i}{A =  \sum_i A_i}

\deqn{B_i =  \psi(O_i, \theta)\psi(O_i, \theta)^T}{B_i = outer(\psi(O_i, \theta), \psi(O_i, \theta))}

\deqn{B =  \sum_i B_i}{B =  \sum_i B_i}

where all of the above are evaluated at \eqn{\hat{\theta}}{hat(\theta)}. The partial derivatives in \eqn{A_i}{A_i}
numerically approximated by the \code{derivFUN}.

Note that \eqn{A =  \sum_i A_i}{A =  \sum_i A_i} and not \eqn{\sum_i A_i/m}{A =  \sum_i A_i/m}, and the same for \eqn{B}{B}.
}
\references{
Stefanski, L. A., & Boos, D. D. (2002). The calculus of m-estimation. The American Statistician, 56(1), 29-38.
}
