% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_matrices}
\alias{compute_matrices}
\title{Compute component matrices for covariance matrix}
\usage{
compute_matrices(geex_list, theta, numDeriv_options = list(method =
  "Richardson"), silent = TRUE, ...)
}
\arguments{
\item{geex_list}{a list containing \code{splitdt} (a \code{data.frame} that
has been \code{\link[base]{split}} by the grouping variable) and \code{eeFUN}
(see \code{\link{estimate_equations}})}

\item{theta}{vector of parameters passed to \code{geex_list$eeFUN}.}

\item{numDeriv_options}{a list of options passed to \code{\link[numDeriv]{jacobian}}.}

\item{...}{additional arguments passed to \code{geex_list$eeFUN}.}
}
\value{
a list with
\itemize{
\item A - the 'bread' matrix
\item B - the 'meat' matrix
\item A_i - the 'bread' matrix for each group
\item B_i - the 'meat' matrix for each group
}
}
\description{
For a given set of estimating equations computes the 'meat' and 'bread' matrices
necessary to compute the covariance matrix.
}

