% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computations.R
\name{compute_sigma}
\alias{compute_sigma}
\title{Compute empirical sandwich covariate estimator}
\usage{
compute_sigma(A, B)
}
\arguments{
\item{A}{a matrix, generally the \eqn{A}{A} matrix returned in the list of matrices from
\code{\link{compute_matrices}}}

\item{B}{a matrix, generally the \eqn{B}{B} matrix returned in the list of matrices from
\code{\link{compute_matrices}}}
}
\description{
Computes \eqn{\Sigma = A^{-1} B (A^{-1})^T }{\Sigma = A^{-1} B (A^{-1})^T} with
provided \eqn{A}{A} and \eqn{B}{B} matrices.
}
