% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computations.R
\name{compute_sigma}
\alias{compute_sigma}
\title{Compute covariance matrix for set of estimating equations}
\usage{
compute_sigma(A, B)
}
\arguments{
\item{A}{the 'bread' matrix returned in the list of matrices from
\code{\link{compute_matrices}}}

\item{B}{the 'meat' matrix returned in the list of matrices from
\code{\link{compute_matrices}}}
}
\description{
Compute covariance matrix for set of estimating equations
}

