% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computations.R
\name{compute_sigma}
\alias{compute_sigma}
\title{Compute covariance matrix for set of estimating equations}
\usage{
compute_sigma(A, B)
}
\arguments{
\item{matrices}{the list of matrices returned from \code{\link{compute_matrices}}}

\item{corrections}{character vector of corrections to perform}
}
\description{
Compute covariance matrix for set of estimating equations
}

