% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_by_funs.R
\name{correct_by}
\alias{correct_by}
\title{Correct sandwich components}
\usage{
correct_by(.components, .correct_control)
}
\arguments{
\item{.components}{an object of class \code{\linkS4class{sandwich_components}}}

\item{.correct_control}{an object of class \code{\linkS4class{correct_control}}}
}
\value{
the result of \code{.FUN} in \code{.correct_control}.
}
\description{
Correct sandwich components
}
