% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_by_funs.R
\name{correction}
\alias{correction}
\title{Creates a correct_control object}
\usage{
correction(FUN, ...)
}
\arguments{
\item{FUN}{a correction to perform. \code{components} must be the
first argument}

\item{...}{additional arguments passed to \code{FUN}}
}
\value{
a \code{\linkS4class{correct_control}} object
}
\description{
Creates a correct_control object
}
\examples{
correction(FUN = fay_bias_correction, b = 0.75)
}
