% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_by_funs.R
\name{correction}
\alias{correction}
\title{Creates a correct_control object}
\usage{
correction(correctFUN, correctFUN_options)
}
\arguments{
\item{correctFUN}{a correction to perform. \code{components} must be the
first argument}

\item{correctFUN_options}{a list of options passed to \code{correctFUN}}
}
\value{
a \code{\linkS4class{correct_control}} object
}
\description{
Creates a correct_control object
}
