% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_funs.R
\name{create_basis}
\alias{create_basis}
\title{Creates an m_estimation_basis object}
\usage{
create_basis(estFUN, data, units, outer_args, inner_args, split_data = NULL)
}
\arguments{
\item{data}{a data.frame}

\item{units}{an optional character string identifying the grouping variable in \code{data}}

\item{split_data}{a \code{split} data.frame}
}
\description{
Creates an m_estimation_basis object
}
