% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_funs.R
\docType{methods}
\name{create_GFUN}
\alias{create_GFUN}
\alias{create_GFUN,m_estimation_basis-method}
\alias{create_GFUN,m_estimation_basis,m_estimation_basis-method}
\title{Creates a function that sums over psi functions}
\usage{
create_GFUN(object, ...)

\S4method{create_GFUN}{m_estimation_basis}(object)
}
\arguments{
\item{object}{an object of class \code{\linkS4class{m_estimation_basis}}}

\item{...}{additional arguments passed to other methods}
}
\description{
From a list of \eqn{\psi(O_i, \theta)}{\psi(O_i, \theta)} for i = 1, ..., m,
creates \eqn{G_m = \sum_i \psi(O_i, \theta)}{G_m = \sum_i \psi(O_i, \theta)},
called \code{GFUN}. Here, \eqn{\psi(O_i, \theta)}{\psi(O_i, \theta)} is the
*inner* part of an \code{estFUN}, in that the data is fixed and \eqn{G_m}{G_m}
is a function of \eqn{\theta)}{\theta}.
}
