% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computations.R
\name{create_GFUN}
\alias{create_GFUN}
\title{Creates a function that sums over psi functions}
\usage{
create_GFUN(psi_list, inner_eeargs = NULL)
}
\arguments{
\item{psi_list}{list of psi functions}

\item{inner_eeargs}{list of arguments passed to psi}
}
\description{
Creates a function that sums over psi functions
}

