% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computations.R
\name{create_psi}
\alias{create_psi}
\title{Creates list of psi functions}
\usage{
create_psi(splitdt, eeFUN, ...)
}
\arguments{
\item{splitdt}{list of dataframes with data per unit}

\item{eeFUN}{the estimating equation function}
}
\description{
Creates list of psi functions
}
