% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_funs.R
\name{create_psi}
\alias{create_psi}
\title{Creates list of psi functions}
\usage{
create_psi(.basis, .approx_control)
}
\arguments{
\item{.basis}{an object of class \code{\linkS4class{m_estimation_basis}}}

\item{.approx_control}{an object of class \code{\linkS4class{approx_control}}
or \code{NULL}}
}
\value{
a list of functions, each function corresponding to a single unit
}
\description{
Creates the estimating function (\eqn{\psi(O_i, \theta)}{\psi(O_i, \theta)})
for each unit. That is, this function evaluates the outer function in
\code{estFUN} for each independent unit and a returns the inner function in
\code{estFUN}.
}
