% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_funs.R
\name{create_psi}
\alias{create_psi}
\title{Creates list of psi functions}
\usage{
create_psi(splitdt, eeFUN, approxFUN = NULL, approxFUN_control = NULL,
  outer_eeargs = NULL)
}
\arguments{
\item{splitdt}{list of dataframes with data per unit}

\item{eeFUN}{the estimating equation function}

\item{approxFUN}{a function that approximates the inner function of \code{eeFUN}.
(EXPERIMENTAL).}

\item{approxFUN_control}{arguments passed to \code{approxFUN}}

\item{outer_eeargs}{a list of arguments passed to \code{eeFUN}}
}
\value{
a list of functions, each function corresponding to a single unit
}
\description{
Creates the estimating function (\eqn{\psi(O_i, \theta)}{\psi(O_i, \theta)})
for each unit. That is, this function evaluates the outer function in \code{eeFUN}
for each independent unit and a returns the inner function in \code{eeFUN}.
}
