% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrections.R
\name{df_correction_2}
\alias{df_correction_2}
\title{Estimate Fay's degrees of freedom correction 2}
\usage{
df_correction_2(A, A_i, C, L, Bbc)
}
\arguments{
\item{A}{the outer "bread" matrix}

\item{C}{same as the C matrix in the Fay (2001) notation Section 2.3}

\item{L}{a k x p matrix where p is the dimension of theta}

\item{Bbc}{the bias corrected "B" matrix}
}
\description{
Estimate Fay's degrees of freedom correction 2
}
