% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrections.R
\name{df_correction_2}
\alias{df_correction_2}
\title{Estimate Fay's degrees of freedom correction 2}
\usage{
df_correction_2(A, A_i, C, L, Bbc)
}
\description{
Estimate Fay's degrees of freedom correction 2
}
