% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrections.R
\name{df_correction_prep}
\alias{df_correction_prep}
\title{Preparations for Fay's degrees of freedom corrections}
\usage{
df_correction_prep(L, A, A_i, H_i)
}
\arguments{
\item{L}{a k x p matrix where p is the dimension of theta}

\item{A}{the outer "bread" matrix}

\item{A_i}{a list of bread matrices per group}

\item{H_i}{a list of bias adjusted matrices}
}
\description{
Preparations for Fay's degrees of freedom corrections
}
