% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{estimate_equations}
\alias{estimate_equations}
\title{ee}
\usage{
estimate_equations(eeFUN, data, units, corrections = NULL,
  correction_options = list(), numDeriv_options = list(method =
  "Richardson"), rootsolver = rootSolve::multiroot,
  rootsolver_options = NULL, findroots = TRUE, roots = NULL, ...)
}
\description{
ee
}

