% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computations.R
\name{estimate_equations}
\alias{estimate_equations}
\title{Estimate parameters and their covariance from a set of estimating equations}
\usage{
estimate_equations(eeFUN, data, units = NULL, corrections_list = NULL,
  numDeriv_options = list(method = "Richardson"),
  rootsolver = rootSolve::multiroot, rootsolver_options = NULL,
  roots = NULL, ee_args = NULL, compute_roots = TRUE,
  compute_vcov = TRUE, ...)
}
\arguments{
\item{eeFUN}{a function that takes in group-level data and returns a function
that takes parameters as its first argument}

\item{data}{a data.frame}

\item{units}{a string identifying the grouping variable in \code{data}}

\item{corrections_list}{an optional list of small sample corrections where each
list element is a list with two elements: `fun` and `options`. See details.}

\item{numDeriv_options}{a list of options for \code{\link[numDeriv]{jacobian}}}

\item{rootsolver}{the function that will find roots of the estimating equations
when \code{compute_roots = TRUE}.}

\item{rootsolver_options}{a list of options for the \code{rootsolver} function}

\item{roots}{a numeric vector containing either starting values for the roots when using
the default \code{rootsolver} or roots that have been estimated elsewhere}

\item{ee_args}{an optional list of arguments passed to the inner function of the `eeFUN`.
See details for an example.}

\item{compute_roots}{whether or not to find the roots of the estimating equations.
Defaults to \code{TRUE}.}

\item{...}{additional arguments passed to the \code{eeFUN}. See details.}
}
\value{
a list with the following
\itemize{
\item \code{parameters} - a vector of estimated parameters
\item \code{vcov} - the variance-covariance matrix for the parameters
\item \code{corrections} - a list of corrected variance-covariance matrices
}
}
\description{
Estimate parameters and their covariance from a set of estimating equations
}
\details{

}
\section{eeFUN arguments}{


Additional arguments may be passed to both the inner and outer function of the `eeFUN`.
Any arguments in `...` are passed to the outer function; any elements of the `ee_args` list
are passed to the inner function. For example, a practical example might be computing a
counterfactual mean using an IPW estimator:

\preformatted{
myeeFUN <- function(data, model){
  X <- model.matrix(model, data = data) #covariates
  A <- data$A #treatment
  Y <- data$Y #outcome
  p <- ncol(X) #number of parameters in model
  function(theta, a){
    Y * (A == a) * 1/plogis(X \%*\% theta[p - 1]) - theta[p]
    # Here theta[p] is the target parameter.
  }
}
}

Then to estimate the mean where `a == 1`:

\preformatted{
estimate_equations(
  eeFUN = myeeFUN,
  data  = mydata,
  units = myunits,
  ee_args = list(a = 1),
  model = mymodel
)
}
}

