% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrections.R
\name{fay_bias_correction}
\alias{fay_bias_correction}
\title{Estimate Fay's bias correction}
\usage{
fay_bias_correction(A_i, A, B_i, B, b = 0.75)
}
\arguments{
\item{A}{the bread matrix}

\item{b}{a numeric value < 1. Defaults to 0.75 as in Fay}

\item{Ai}{a list of bread matrices per group}

\item{Bi}{a list of meat matrices per group}
}
\value{
a corrected covariance matrix
}
\description{
Estimate Fay's bias correction
}

