% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrections.R
\name{fay_bias_correction_partial}
\alias{fay_bias_correction_partial}
\title{Compute the matrices necessary for Fay's bias correction}
\usage{
fay_bias_correction_partial(A, A_i, B, B_i, b)
}
\arguments{
\item{A}{the bread matrix}

\item{b}{a numeric value < 1. Defaults to 0.75 as in Fay}

\item{Ai}{a list of bread matrices per group}

\item{Bi}{a list of meat matrices per group}
}
\value{
a list with
\itemize{
\item H_i - the H_i matrix
\item Bbc - the 'bias corrected' meat matrix
}
}
\description{
Compute the matrices necessary for Fay's bias correction
}

