% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrections.R
\name{fay_df_correction}
\alias{fay_df_correction}
\title{Fay's degrees of freedom corrections}
\usage{
fay_df_correction(A_i, A, B_i, B, b = 0.75, L, version)
}
\description{
Fay's degrees of freedom corrections
}

