% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{geex-summary-class}
\alias{geex-summary-class}
\title{geex summary object}
\description{
geex summary object
}
\section{Slots}{

\describe{
\item{\code{basis}}{a \code{\linkS4class{m_estimation_basis}} object}

\item{\code{corrections}}{a \code{list} of correction performed on \code{sandwich_components}}

\item{\code{estimates}}{a \code{numeric} vector of parameter estimates}

\item{\code{vcov}}{the empirical sandwich variance \code{matrix}}
}}

