% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{geex_control-class}
\alias{geex_control-class}
\title{geex_control S4 class}
\description{
A general class for defining a \code{function}, and the options passed to the
function
}
\section{Slots}{

\describe{
\item{\code{.FUN}}{a function}

\item{\code{.options}}{a list of options passed to \code{.FUN}}
}}

\seealso{
root_control deriv_control approx_control
}
