% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{grab}
\alias{grab}
\title{Grab something from an object}
\usage{
grab(from, what, ...)
}
\arguments{
\item{from}{an object}

\item{what}{what to grab}

\item{...}{additional arguments passed to grab function}
}
\description{
Grab something from an object
}
