% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab.R
\name{grab}
\alias{grab}
\title{Grab something from an object}
\usage{
grab(from, what, ...)
}
\arguments{
\item{from}{an object}

\item{what}{what to grab one of 'response', 'design_matrix', 'response_formula',
'fixed_formula', 'eeFUN'}

\item{...}{additional arguments passed to \code{grab_**} function}
}
\description{
Grab something from an object
}
