% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{grab_bread}
\alias{grab_bread}
\alias{grab_bread,sandwich_components-method}
\alias{grab_bread,sandwich_components,sandwich_components-method}
\title{Grabs the .A (bread matrix) slot}
\usage{
grab_bread(object)

\S4method{grab_bread}{sandwich_components}(object)
}
\arguments{
\item{object}{a \code{\linkS4class{sandwich_components}} object}
}
\description{
Grabs the .A (bread matrix) slot
}
