% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{grab_meat}
\alias{grab_meat}
\alias{grab_meat,sandwich_components-method}
\alias{grab_meat,sandwich_components,sandwich_components-method}
\title{Gets the .B (meat matrix) slot}
\usage{
grab_meat(object)

\S4method{grab_meat}{sandwich_components}(object)
}
\arguments{
\item{object}{a \code{\linkS4class{sandwich_components}} object}
}
\description{
Gets the .B (meat matrix) slot
}
\examples{
myee <- function(data) {
  function(theta) {
    c(
      data$Y1 - theta[1],
      (data$Y1 - theta[1])^2 - theta[2]
    )
  }
}

results <- m_estimate(
  estFUN = myee,
  data = geexex,
  root_control = setup_root_control(start = c(1, 1))
)

grab_meat_list(results@sandwich_components)
}
