% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_funs.R
\name{grab_response_formula}
\alias{grab_response_formula}
\title{Grab the LHS formula from a model object}
\usage{
grab_response_formula(model)
}
\arguments{
\item{model}{a model object such as \code{lm}, \code{glm}, \code{merMod}}
}
\value{
the left-hand side of a model's \code{\link[stats]{formula}} object
}
\description{
Grab the LHS formula from a model object
}
\examples{
fit <- lm(Sepal.Width ~ Petal.Width, data = iris)
grab_response_formula(fit)
}
