% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{m_estimation_basis-class}
\alias{m_estimation_basis-class}
\title{m_estimation_basis S4 class}
\description{
m_estimation_basis S4 class
}
\section{Slots}{

\describe{
\item{\code{.data}}{the analysis data.frame}

\item{\code{.units}}{an (optional) character string identifying the variable in
\code{.data} which splits the data into indepedent units}

\item{\code{.split_data}}{list formed by \code{split(.data, f = .data[[.units]])}
(optional). This is automatically created if not provided.}

\item{\code{.weights}}{a numeric vector of weights used in weighting the estimating
functions}

\item{\code{.psiFUN_list}}{a list of \code{psiFUN}s created by \code{\link{create_psiFUN_list}}}
}}

