% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrections.R
\name{make_corrections}
\alias{make_corrections}
\title{Make corrections to basic sandwich estimators}
\usage{
make_corrections(mats, corrections)
}
\arguments{
\item{mats}{list of matrices resulting from \code{\link{compute_matrices}}}

\item{corrections}{a list with one sublist  for each correction to make. The
sublist must contain `fun` and `options` arguments}
}
\value{
a list with corrected values
}
\description{
Make corrections to basic sandwich estimators
}
