% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_functions.R
\name{make_eefun.merMod}
\alias{make_eefun.merMod}
\title{glmer Estimating Equations}
\usage{
\method{make_eefun}{merMod}(model, data, numderiv_opts = NULL, ...)
}
\arguments{
\item{model}{a model object object}

\item{data}{data with which to create the estimating equation function}

\item{numderiv_opts}{a list of argument passed to \link{\code{numDeriv::grad}}}

\item{...}{passed to methods}
}
\description{
Create estimating equation function from a \code{merMod} object
}
