% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_functions.R
\name{objFun_merMod}
\alias{objFun_merMod}
\title{glmer Objective Fundtion}
\usage{
objFun_merMod(family, ...)
}
\arguments{
\item{family}{distribution family of objective function}

\item{...}{additional arguments pass to objective function}
}
\description{
glmer Objective Fundtion
}
