% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_eeFUN_funs.R
\name{objFun_merMod}
\alias{objFun_merMod}
\title{Objective Function for merMod object}
\usage{
objFun_merMod(family, ...)
}
\arguments{
\item{family}{distribution family of objective function}

\item{...}{additional arguments pass to objective function}
}
\description{
Objective Function for merMod object
}
