% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,sandwich_components-method}
\alias{show,sandwich_components,sandwich_components-method}
\alias{show,m_estimation_basis-method}
\alias{show,m_estimation_basis,m_estimation_basis-method}
\alias{show,geex-method}
\alias{show,geex,geex-method}
\alias{show,geex_summary-method}
\alias{show,geex_summary,geex_summary-method}
\title{Show (print) the S4 geex classes}
\usage{
show(object)

\S4method{show}{sandwich_components}(object)

\S4method{show}{m_estimation_basis}(object)

\S4method{show}{geex}(object)

\S4method{show}{geex_summary}(object)
}
\arguments{
\item{object}{the object to print}
}
\description{
Show (print) the S4 geex classes

\code{\linkS4class{m_estimation_basis}}, or \code{\linkS4class{geex}} object

Shows the m_estimation_basis

Shows the geex object

Shows the geex_summary object
}
