% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{show,sandwich_components-method}
\alias{show,sandwich_components-method}
\alias{show,sandwich_components,sandwich_components-method}
\alias{show,m_estimation_basis-method}
\alias{show,m_estimation_basis,m_estimation_basis-method}
\alias{show,geex-method}
\alias{show,geex,geex-method}
\alias{show,geex-summary-method}
\alias{show,geex-summary,geex-summary-method}
\title{Shows the sandwich_components S4 class}
\usage{
\S4method{show}{sandwich_components}(object)

\S4method{show}{m_estimation_basis}(object)

\S4method{show}{geex}(object)

\S4method{show}{`geex-summary`}(object)
}
\arguments{
\item{object}{a \code{\linkS4class{sandwich_components}},
\code{\linkS4class{m_estimation_basis}}, or \code{\linkS4class{geex}} object}
}
\description{
Shows the sandwich_components S4 class

Shows the m_estimation_basis

Shows the geex object

Shows the geex-summary object
}
