% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{show,sandwich_components-method}
\alias{show,sandwich_components-method}
\title{Shows the sandwich_components}
\usage{
\S4method{show}{sandwich_components}(object)
}
\description{
Shows the sandwich_components
}
